requirements.txt
matplotlib==3.3.3
numpy==1.19.3
Pillow==8.0.1
PyQt5==5.15.1
Keras==2.2.4
tensorflow==1.8.0